refactoring : 서비스 구조 수정
