+++ src/main/java/com/runninghi/RunninghiApplication.java
@@ -1,7 +1,6 @@
 package com.runninghi;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
