+++ src/main/java/com/runninghi/feedback/command/domain/service/FeedbackDomainService.java
@@ -1,0 +1,5 @@
+package com.runninghi.feedback.command.domain.service;
+
+public interface FeedbackDomainService {
+
+}
