+++ src/main/java/com/runninghi/feedback/command/infrastructure/service/FeedbackInfraService.java
@@ -1,0 +1,6 @@
+package com.runninghi.feedback.command.infrastructure.service;
+
+import com.runninghi.feedback.command.domain.service.FeedbackDomainService;
+
+public class FeedbackInfraService implements FeedbackDomainService {
+}
