--- src/main/java/com/runninghi/feedback/command/infrastructure/service/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.feedback.command.infrastructure.service;
-
-public class test {
-}
